+++ apps/web/pages/apps/[appId]/pages/index.tsx
@@ -2,7 +2,7 @@ import { withPageAuthRequired } from '@auth0/nextjs-auth0'
 import { NextPageTemplate } from '@codelab/frontend/abstract/props'
 import { MainPanePage } from '@codelab/frontend/modules/page'
 import { withAppQueryProvider } from '@codelab/frontend/presenter/container'
-import { DashboardLayout } from 'apps/web/src/layout/DashboardLayout'
+import { DashboardLayout } from '@codelab/frontend/view/templates'
 import React from 'react'
 
 const Pages: NextPageTemplate<'dashboard'> = () => <></>
